Refactor parallel runner tests
